+++ src/main/java/com/fiap/restaurante/infrastructure/adapter/in/request/PedidoRequest.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import com.fiap.restaurante.core.domain.Pedido;
 
-public record PedidoRequest(Integer idCliente, List<PedidoProdutoRequest> listaPedidoProdutos) {
+public record PedidoRequest(Long idCliente, List<PedidoProdutoRequest> listaPedidoProdutos) {
     
     public Pedido toDomain(){
         return new Pedido(idCliente, listaPedidoProdutos.stream().map(pedidoProdutos -> pedidoProdutos.toDomain()).toList());
